- 修改细节

+++ src/main/java/com/centerm/fud_demo/entity/User.java
@@ -13,7 +13,7 @@ public class User {
     private String createTime;
     private Integer state;
     private Integer isOnline;
-    private Long roleId;
+    private Integer roleId;
     //salt 为用户名
 
     public User() {
